devices: virtio-blk: fix GET_ID behavior after reset
The activate() function would originally take self.id and not restore it
during reset. The simplest way to fix this is to just copy the id rather
than taking it, which should not be a big deal since it is small (20
bytes).

Fixes virtio-blk id support across device resets, such as when booting
with UEFI firmware.

BUG=None
TEST=cat /sys/block/vda/serial

Change-Id: I29666cae34fccc87668a2a56f155eb5e387eaed0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4522148
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: James Ye <jamesye@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>