sandbox: Call lower_token before run_control
Upstreaming a fix:

We found that token lowering can conflict with valid operations if they
happen in another thread, concurrently. An example is creating a named
pipe / Tube pair, where the client->server connection fails with access
denied if the token is lowered in between the two operations.

We fix this for now, by moving lower_token before run_control with a big
warning of the pitfalls of moving it again.

BUG=b:281108137
TEST=Ran emulator

Change-Id: I1bce32edc6872f895630eb3005427521e3653228
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4544898
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Idan Raiter <idanr@google.com>