devices: virtio: add Reader::peek_obj() and remove Clone
This makes it possible to read an object from a Reader without consuming
it, which can be used to replace a sequence like .clone().read_obj().
Therefore Clone is no longer needed on Reader (and Writer already did
not need it), so the #[derive(Clone)] can be removed.

The core functionality of the io::Read implementation for Reader is
factored out into a new copy_regions_to_mut_ptr() function so it can be
reused by both peek_obj() and io::Read.

BUG=None
TEST=cargo test -p devices reader_peek_obj
TEST=tools/dev_container tools/presubmit

Change-Id: I8cca912a34ced46bbd0c2d75a84f3eca65fe6b56
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4545950
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>