devices: pci: provide bar_index and offset in read/write_bar()
The PciDevice::read_bar() and PciDevice::write_bar() functions are now
passed a BAR index and offset within that BAR, rather than a guest
physical address. This simplifies the job of each device implementation
since it is no longer necessary to look up which BAR corresponds to the
provided address (this is handled in generic PciDevice code now).

Providing the address information in this way also ensures devices will
work even in the face of BAR reprogramming; there is no way to compare
against a stale BAR address, since that information is not provided to
the read_bar()/write_bar() callbacks anymore.

BUG=None
TEST=tools/dev_container tools/presubmit
TEST=Boot Crostini and ARCVM on brya

Change-Id: Ia200c82ebfb58a5ae1e8015fb58fd9bbc34cce63
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4564155
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>