src: no-op ack IRQ token refresh requests.
IRQ token refreshes are needed for restoring a VM with a different shape
than the one originally booted. We don't need this feature right now,
but we do need to let the generic code that expects this to work that we
have "completed" the request.

BUG=b:283520036
TEST=ran emulator, snapshotted & restored on Windows.

Change-Id: Ie7c9a6df3de55907e5a2d57afafb89afa91fce01
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4571993
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Elie Kheirallah <khei@google.com>
Reviewed-by: Richard Zhang <rizhang@google.com>