devices: net: move vhost-net device path to NetParameters
This is the last network-related argument that is not part of
NetParameters, so move it there.

It is also required as we want to be able to instanciate a network
device from NetParameters only, and not the whole configuration.

BUG=b:217480278
TEST=presubmit

Change-Id: I39dcac26569ff323c8afb5c12ef183f313db2862
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4583706
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>