devices: virtio: Remove VirtioDeviceSaved and cleanup
Change of design - queues will be saved in the virtio device layer as
opposed to being handed back to the VirtioPciDevice layer. As such, this
struct is no longer required, and the stop() function is no longer
necessary as we will not be passing back ownership of the queues.
Moved stop() implementations to virtio_sleep for devices that had it
implemented.

BUG=b:232437513
TEST=presubmit

Change-Id: I569b5cb645e5e11a2467757718c41ef2e77f7921
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4583464
Reviewed-by: Richard Zhang <rizhang@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>