vhost-user: gpu: Allow queue stop without panic
The state variable holds onto a queue internally, if we reset it we can
drop the refcount to zero and release the remaining resources.

Also, gate the stop_queue behind if the queue is set, since the cursor
queue is not used and will always give an error.

TEST=downstream can shutdown cleanly
BUG=b:284156457

Change-Id: If9fd56cb5aeb94456726989248c40d6cb9e006df
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4602368
Reviewed-by: Richard Zhang <rizhang@google.com>
Commit-Queue: Idan Raiter <idanr@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>