devices: virtio: snapshot-restore for VirtioPciDevice
`VirtioDevice` specific methods are added for sleep, wake, snapshot, and
restore lifecycle instead of using the `Suspendable` trait. The main
difference is that ownership of the virtio queues is passed around as
part of sleep and wake so that `VirtioPciDevice` can handle more of the
details.

If this is accepted, we should remove `stop` and `Suspendable` from
`VirtioDevice` because they are redundant with the new methods.

This change borrows code from rizhang@'s https://crrev.com/c/4550457/14.
It also diverges from the approach in that CL. It might be that this
approach isn't feasible for vhost-user devices and we'll need to redo
in the near future, but this does work for normal virtio devices and
I've successfully tested end-to-end with virtio-block.

There is duplicate code between `activate` and `restore` that is worth
deduping. I'm holding off for now so that we don't "overfit" the code to
this solution before figuring out the vhost-user half.

BUG=b:266514327,b:266623213

Change-Id: If3b248834cefcf2f09f875259a35168a86292b9b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4602373
Reviewed-by: Elie Kheirallah <khei@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>