devices: console: Change Console to be event driven
Console currently runs using a detached thread with a blocking read.
This CL changes the implementation to be Event driven, by getting a
read notifier. This allows us to also gain control over the input thread
and allows joining the thread when shutting down the console.

Split off implementation of function `spawn_input_thread` between
windows and unix. Windows uses PipeConnection which doesn't implement
ReadNotifier, whereas unix does. This is needed to change the
implementation of Console for it to be event driven as opposed to having
blocking reads in detached threads.

BUG=b:266514274
Test=presubmit

Change-Id: Id98a611e7170ec7a1dd0c94fde0be7f6bb0f5b44
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4607766
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Richard Zhang <rizhang@google.com>