devices: virtio_pci_device: restore fix
Restore right now tries to restore interrupts at all times. If the
device was not previously activated, the interrupt will always be null
which will cause the restore to always fail. Add a check to see if the
device was previously activated before restoring the interrupt.

BUG=N/A
Test=presubmit

Change-Id: Ifddc27f2a63bab7b1509b934a9c14d129977a401
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4610691
Reviewed-by: Richard Zhang <rizhang@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>