e2e_tests: Fix snapshot/restore tests
Following this CL (https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4602373),
attempting a restore while a VM is running will cause the restore to
fail and the tests to crash as a result.
Change the tests to start a new VM with the previous snapshot config to
validate that the VM restored properly.
Enable suspend/resume tests.

BUG=N/A
TEST=presubmit + e2e_tests

Change-Id: Ifdd63fee7953cd1b7be208a093802da388322c81
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4616833
Reviewed-by: Noah Gold <nkgold@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Richard Zhang <rizhang@google.com>