devices: make virtio queues optional in snapshotting.
Currently every queue must be used by a device for it to snapshot, since
there is no optionality in the returned set of queues (see
virtio_sleep). This CL modifies the typing so queues are optional by
storing both the queue and its index.

BUG=b:266514168
TEST=builds
TEST=snapshotted & restored a minimal VM with these changes (note: we
don't have a test VM with a disk yet, so we don't have full coverage
here).

Change-Id: I05f70633861f868dac1faeaddf86921e7684a008
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4618962
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Richard Zhang <rizhang@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>