unix: vcpu: use thread_local try_with() in signal handler
The thread_local LocalKey::with() function can panic if it is called
while the thread local is being destroyed[1], which could happen if the
VCPU thread gets a signal while it is exiting. Use `try_with()` instead
to prevent the panic in this case, and just discard any errors, since if
the thread is already exiting, the VCPU does not need to be notified.

[1]: https://doc.rust-lang.org/std/thread/struct.LocalKey.html#method.with

BUG=b:286786420
TEST=Start and stop Crostini on brya

Change-Id: I1dab03c495bc7cffb85234fabe51253e456a83c4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4628139
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>