devices: Retain device status after snapshot/restore.
When doing a snapshot/restore, Vcpus and Devices are made sure
to be asleep. The device controller gets the initial Vcpu state so that
Vcpus return to their prior state after the snapshot/restore
operations. Devices currently do not have the functionality of remaining
in the same state and will always go back to `running` at the end of the
operation. Adding a state save to devices.

BUG=N/A
TEST=presubmit, e2e_tests

Change-Id: Ibfb6c4055cb1f852ec6a85043b6464c621a83d59
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4641774
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>