devices: wake every virtio doorbell on restore.
When we snapshot a virtio device, there may be data in the avail
ring of the virtqueue. (We do typically flush these queues by
freezing VCPUs and giving bias to devices emptying their queues
before sleeping, but it is not guaranteed for all devices.) In
this situation, the driver may not notify the device again, which
will leave the device stalled.

To avoid that situation, we can explicitly signal all virtqueue
doorbells on resume. This will be a spurious wakeup for the
device, but spurious wakeups are safe (in that case, the device
will find no descriptors to process and go back to sleep).

BUG=b:283017441
TEST=ran e2e tests; manually snapshotted & restored a minimal VM.

Change-Id: I930c3cd9e6745e6e88fcd69bce4aff7534568e9c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4690997
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>